Fix user fields
FacebookRequest excepts the fourth parameter to be an associative array
with one of the keys being `fields`. The value must be comma separated
list of fields.

Not sure if (and at which point) this has been changed but #22 seems to
be the one that first introduced `['fields' => []]`. Then #23 broke
that. So this is to fix #23 and to keep the fields configurable.